refactor: consolidate serializer utilities
