+++ advanced_alchemy/service/_sync.py
@@ -29,7 +29,7 @@ from advanced_alchemy.repository.typing import MISSING, ModelT, OrderingPair, Pr
 from advanced_alchemy.service._util import ResultConverter
 from advanced_alchemy.utils.dataclass import Empty, EmptyType
 from advanced_alchemy.utils.deprecation import warn_deprecation
-from advanced_alchemy.utils.serializers import (
+from advanced_alchemy.utils.serialization import (
     UNSET,
     BulkModelDictT,
     ModelDictListT,
