+++ tests/fixtures/uuid/services.py
@@ -6,7 +6,7 @@ from advanced_alchemy.service import (
     SQLAlchemyAsyncRepositoryService,
     SQLAlchemySyncRepositoryService,
 )
-from advanced_alchemy.utils.serializers import (
+from advanced_alchemy.utils.serialization import (
     ModelDictT,
     is_dict_with_field,
     is_dict_without_field,
