+++ tests/unit/test_attrs_integration.py
@@ -7,7 +7,7 @@ from typing import Any, Optional
 import pytest
 from attrs import define, field
 
-from advanced_alchemy.utils.serializers import (
+from advanced_alchemy.utils.serialization import (
     ATTRS_INSTALLED,
     AttrsInstance,
     is_attrs_instance,
