+++ tests/unit/test_file_object.py
@@ -13,7 +13,7 @@ from sqlalchemy.orm import Session
 from advanced_alchemy.types.file_object import FileObject
 from advanced_alchemy.types.file_object.base import StorageBackend
 from advanced_alchemy.types.file_object.session_tracker import FileObjectSessionTracker
-from advanced_alchemy.utils.serializers import PYDANTIC_INSTALLED, BaseModel
+from advanced_alchemy.utils.serialization import PYDANTIC_INSTALLED, BaseModel
 
 if sys.version_info >= (3, 11):
     from builtins import ExceptionGroup
